+++ src/Manifest.ts
@@ -1,9 +1,6 @@
 import Store from "./Store";
+import BookResourceStore from "./BookResourceStore";
 import * as Utils from "./Utils";
-
-type xmlObject = {
-  [key: string]: string[] | string | xmlObject | [];
-};
 
 export interface Link {
   rel?: Array<string>;
@@ -20,47 +17,6 @@ export interface Metadata {
   identifier?: string;
   language?: string;
   modified?: string;
-}
-
-function xmlToJson(xml: any) {
-  let obj = {} as xmlObject;
-
-  // process ELEMENT_NODE
-  if (xml.nodeType == 1) {
-    if (xml.attributes.length > 0) {
-      obj["@attributes"] = {};
-      for (let j = 0; j < xml.attributes.length; j++) {
-        const attribute = xml.attributes.item(j);
-        obj["@attributes"][attribute.nodeName] = attribute.nodeValue;
-      }
-    }
-  }
-  // process TEXT_NODE
-  else if (xml.nodeType == 3) {
-    obj = xml.nodeValue;
-  }
-
-  if (xml.hasChildNodes()) {
-    for (let i = 0; i < xml.childNodes.length; i++) {
-      const item = xml.childNodes.item(i);
-      const nodeName: string = item.nodeName;
-
-      if (typeof obj[nodeName] == "undefined") {
-        obj[nodeName] = xmlToJson(item);
-      } else {
-        //@ts-ignore
-        if (typeof obj[nodeName].push == "undefined") {
-          const old = obj[nodeName];
-          obj[nodeName] = [];
-          //@ts-ignore
-          obj[nodeName].push(old);
-        }
-        //@ts-ignore
-        obj[nodeName].push(xmlToJson(item));
-      }
-    }
-  }
-  return obj;
 }
 
 function parseOPFPackage(OPFPackage: any): any {
@@ -89,18 +45,19 @@ function parseOPFResources(OPFPackage: any): any {
   );
 }
 
-function base64encodeImage(fullResourceUrl: string) {
-  let baseImage = new Image();
-  baseImage.setAttribute("crossOrigin", "anonymous");
-  baseImage.src = fullResourceUrl;
+// function base64encodeImage(fullResourceUrl: string) {
+//   let baseImage = new Image();
+//   baseImage.setAttribute("crossOrigin", "anonymous");
+//   baseImage.src = fullResourceUrl;
 
-  var canvas = document.createElement("canvas");
-  canvas.width = baseImage.width;
-  canvas.height = baseImage.height;
-  var ctx = canvas.getContext("2d");
-  ctx && ctx.drawImage(baseImage, 0, 0);
-  return canvas.toDataURL("image/png");
-}
+//   var canvas = document.createElement("canvas");
+//   canvas.width = baseImage.width;
+//   canvas.height = baseImage.height;
+//   var ctx = canvas.getContext("2d");
+//   ctx && ctx.drawImage(baseImage, 0, 0);
+//   return canvas.toDataURL("image/png");
+// }
+
 /* Manifest is constructed from manifest.json or Package Document */
 export default class Manifest {
   public readonly metadata: Metadata;
@@ -110,10 +67,31 @@ export default class Manifest {
   public readonly toc: Array<Link>;
   private readonly manifestUrl: URL;
 
+  public static async getManifestUrlFromContainer(containerHref: string): Promise<URL> {
+   return window
+      .fetch(containerHref)
+      .then((response) => response.text())
+      .then((text) => {
+        let blah = new window.DOMParser().parseFromString(text, "text/html");
+        return blah;
+      })
+      .then((xml) =>
+        xml.getElementsByTagName("rootfile")[0]
+          ? xml.getElementsByTagName("rootfile")[0].getAttribute("full-path")
+          : ""
+      )
+      .then((rootfile:string) => {
+        console.log("rootFile", rootfile);
+        const url = containerHref.replace("META-INF/container.xml", rootfile);
+        return rootfile ? new URL(url) : new URL("");
+      });
+  }
+
   /* Fetch Package Document (OEBPS package file) or Webpub Manifest (manifest.json)*/
   public static async getManifest(
     manifestUrl: URL,
-    store?: Store
+    store?: Store,
+    resourceStore?: BookResourceStore
   ): Promise<Manifest> {
     const fetchManifest = async (): Promise<Manifest> => {
       const isJSONManifest = Boolean(manifestUrl.href.endsWith(".json"));
@@ -128,35 +106,35 @@ export default class Manifest {
             .then((str) =>
               new window.DOMParser().parseFromString(str, "text/xml")
             )
-            .then((data) => JSON.stringify(xmlToJson(data)));
+            .then((data) => JSON.stringify(Utils.xmlToJson(data)));
 
-      if (store) {
+      if (resourceStore) {
         const resources = parseOPFResources(parseOPFPackage(manifest));
+        const manifestPath = manifestUrl.href.substring(
+          0,
+          manifestUrl.href.lastIndexOf("/")
+        );
+        console.log("path", manifestPath);
 
-        for (let resource of resources) {
-          const fullResourceUrl = `${manifestUrl.href.replace(
-            /[a-z]+.opf/,
-            ""
-          )}${resource.href}`;
+        await resources.forEach(async (resource: any) => {
+          //For some reason calling resource.href below results in undefined unless it is assigned first
+          let resourceUrl = resource.href;
 
-          let encodedImage;
-
-          const isImage = Boolean(
-            fullResourceUrl.match(/(jpe?g|png|gif|bmp|css)$/)
-          );
+          //add leading slash to resource.href if not exists
+          const fullResourceUrl =
+            resource.href[0] === "/"
+              ? `${manifestPath}${resourceUrl}`
+              : `${manifestPath}/${resourceUrl}`;
 
           /* store each resource in store */
-          if (isImage) {
-            encodedImage = await base64encodeImage(fullResourceUrl);
-            await store.set(resource.href, encodedImage);
-          } else {
-            await window
-              .fetch(fullResourceUrl)
-              .then((response) => response.text())
-              .then((content) => store.set(resource.href, content));
-          }
-        }
+          resource = await fetch(fullResourceUrl);
+          let blob = await resource.blob();
+          let url = URL.createObjectURL(blob);
+          await resourceStore.addBookData(resourceUrl, blob, url);
+        });
+      }
 
+      if (store) {
         await store.set("manifest", JSON.stringify(manifest));
       }
 
@@ -184,7 +162,7 @@ export default class Manifest {
       }
     }
 
-    return fetchManifest();
+    return await fetchManifest();
   }
 
   parseOPFPackage(OPFPackage: any): any {
@@ -225,6 +203,7 @@ export default class Manifest {
       []
     );
   }
+
   public parseOPFSpine(OPFPackage: any): any {
     const emptySpine: string[] = [];
     return (OPFPackage?.package?.spine?.itemref || emptySpine).reduce(